abstract out moveCellDown
